add get klusterlet-info command add common utility PrintComponentsCRD for cluster-manager/klusterlet to print crd
